1223~0102 배경 오브젝트 부모 클래스 추가 & 변수명 규칙 변경 & 프레임 관리 방식 변경 적용 진행 & Etc
배경 오브젝트 부모 클래스 추가
- BackgroundObject 클래스를 생성
  - 해당 클래스를 배경 오브젝트 클래스에 상속시킴.
- 배경 오브젝트들이 공통적으로 가지는 요소들을 추출하여 부모 클래스로 이동
-

변수명 규칙 변경
- Skill, Enemy, Player 등에 카멜 표기법 적용

프레임 관리 방식 변경 적용 진행
- 정적 맴버 함수를 통한 프레임 관리 방식 적용 작업을 진행
- Enemy, Item, UI, Projectile
- Player는 현재 중복 생성될 일이 없어 해당 방식을 적용하지 않음

Etc
- 사용되지 않는 일부 파일 제거 진행
- Text에서 표현 가능한 글자를 추가함
- PickRange 증가 시에 문제가 발생할 여지가 있어 해당 부분 임시 수정.
  - 최대 크기 도달 시 더 이상 증가되지 않도록 예외 처리 진행
  - 추후 최대 크기 도달 시 스킬 선택지에 나타나지 않도록 막을 예정

+++ DX_MyProject/Framework/Collision/CircleCollider.cpp
@@ -3,10 +3,10 @@
 void CircleCollider::CreateLine()
 {
     float angle = (M_PI * 2);
-    for (int i = 0; i <= vertex_count; i++)
+    for (UINT i = 0; i <= vertex_count; i++)
     {
         verticies.emplace_back(cos(angle) * radius, sin(angle) * radius);
-        angle -= 2 * M_PI / vertex_count;
+        angle -= 2 * (float)(M_PI) / vertex_count;
     }
 
     VB = new VertexBuffer(verticies.data(), sizeof(VertexPos), verticies.size());
